ai/llm-chat-server/pyproject.toml
[project]
name = "llm-chat-server"
version = "0.1.0"
description = "A server for chat with LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-storage-blob>=12.25.1",
    "fastapi>=0.115.12",
    "google-genai>=1.8.0",
    "openai>=1.69.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
